// vSweatbox/src/types/common.ts
export type FlightStatus = "ramp" | "clearedIFR" | "pushback" | "taxi" | "departing" | "handedOff" | "handedOffReminded" | "departed";

export type FlightPlan = {
    callsign: string;
    CID: string;
    squawk: string;
    aircraftType: string;
    actualAircraftType: string;
    requestedAltitude?: string;
    equipmentCode: string;
    departure: string;
    destination: string;
    plannedTime: string;
    speed: string;
    size: number;
    altitude: string;
    route: string;
    remarks: string;
    positionX: number;
    positionY: number;
    rotation: number;
    printCount: number;
    routeType: "TEC" | "H" | "pattern" | "VFR" | "VFRFF";
    requests: AircraftRequest[];
    canSendRequestTime: number;
    pushbackLocation: { x: number, y: number};
    status: FlightStatus;
    statusChangedTime?: number;
    created: boolean;
    parkingSpotId: string;
    taxiwayNodeId?: string;
}
// ...
export type RequestReminderType = "readbackIFR" | "taxiVFR" | "aircraftHandoff";

export type AircraftRequest = {
    callsign: string;
    priority: number;
    responseMessage?: string;
    requestMessage?: string;
    subsequentRequest?: AircraftRequest;
    nextRequestDelay: number;
    atcMessage?: string;
    nextStatus?: FlightStatus;
    reminder?: {
        message: string;
        sendDelay: number;
        type: RequestReminderType;
        sendTime?: number;
    }
}